operator/policy: validate endpoint namespace in Sanitize
Move namespace validation from operator's checkEndpointSelectorNamespace
into a new SanitizeWithNamespace method on api.Rule. This ensures both
the operator and agent (via cnp_types.go Parse) use the same validation
code, preventing divergence. CCNP continues to use plain Sanitize() as
it is cluster-scoped.

Bahtya
